agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "agent using crewAI"
authors = [{ name = "Muhammad Hamza", email = "muhammadhamzaali077@gmail.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.100.1,<1.0.0",
    "litellm>=1.59.8",
    "google-generativeai>=0.6.0",
    "google-auth-oauthlib>=1.2.1",
    "python-dateutil>=2.9.0.post0",
    "tzlocal>=5.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
kickoff = "agent.main:kickoff"
start = "agent.main:RoutedFlow"
papa = "agent.papa:main"
agent = "agent.agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
